Update presentation to latest shaderback.js
The presentation was running on an older version. Since the fragment
and vertex shaders had different float precisions set an error was
generated (can't share uniforms with different precisions across
both shaders). Not sure why this used to be okay, but it became
broken at some point (Firefox update). The move to the latest
version, and fixing the precision, fixes the issue.